+++ lnd-grpc-client/lnd-grpc-client-autoconfigure/src/integTest/java/org/tbk/lightning/lnd/grpc/config/DisabledLndHealthIndicatorIntegrationTest.java
@@ -1,6 +1,5 @@
 package org.tbk.lightning.lnd.grpc.config;
 
-import com.google.common.collect.ImmutableList;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,39 +7,37 @@ import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.autoconfigure.web.servlet.MockMvcPrint;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.util.List;
-
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @Slf4j
 @SpringBootTest
-@AutoConfigureMockMvc(print = MockMvcPrint.LOG_DEBUG, printOnlyOnFailure = false)
+@AutoConfigureMockMvc
 @ActiveProfiles("test")
 @TestPropertySource(properties = {
         "server.port=13337",
         "management.server.port=13337",
-        "management.endpoints.web.exposure.include=metrics",
+        "management.endpoint.health.show-details=always",
+        "management.health.lndJsonRpc.enabled=false",
         "org.tbk.lightning.lnd.grpc.rpchost=localhost",
         "org.tbk.lightning.lnd.grpc.rpcport=13337",
         "org.tbk.lightning.lnd.grpc.macaroonFilePath=/dev/null",
         "org.tbk.lightning.lnd.grpc.certFilePath=src/test/resources/lnd/tls-test.cert"
 })
-public class LndJsonRpcClientMetricsIntegrationTest {
+public class DisabledLndHealthIndicatorIntegrationTest {
 
     @SpringBootApplication
-    public static class BitcoinJsonRpcTestApplication {
+    public static class LndJsonRpcTestApplication {
 
         public static void main(String[] args) {
             new SpringApplicationBuilder()
-                    .sources(BitcoinJsonRpcTestApplication.class)
+                    .sources(LndJsonRpcTestApplication.class)
                     .web(WebApplicationType.SERVLET)
                     .run(args);
         }
@@ -50,19 +47,16 @@ public class LndJsonRpcClientMetricsIntegrationTest {
     private MockMvc mockMvc;
 
     @Test
-    public void itShouldAddMetricsEndpoints() throws Exception {
-        List<String> metricNames = ImmutableList.<String>builder()
-                .add("lnd.blocks.height")
-                .add("lnd.channels.active")
-                .add("lnd.channels.inactive")
-                .add("lnd.channels.pending")
-                .add("lnd.peers")
-                .build();
+    public void itShouldCheckHealthEndpointDoesNotExist() throws Exception {
+        mockMvc.perform(get("/actuator/health/lndJsonRpc"))
+                .andExpect(status().isNotFound());
+    }
 
-        for (String metricName : metricNames) {
-            mockMvc.perform(get("/actuator/metrics/{metricName}", metricName))
-                    .andDo(print())
-                    .andExpect(status().isOk());
-        }
+    @Test
+    public void itShouldNotAddHiddenServiceInformationToHealthEndpoint() throws Exception {
+        mockMvc.perform(get("/actuator/health"))
+                .andExpect(jsonPath("status").value("UP"))
+                .andExpect(jsonPath("components.lndJsonRpc").doesNotExist())
+                .andExpect(status().isOk());
     }
 }