+++ lnd-grpc-client/lnd-grpc-client-autoconfigure/src/test/java/org/tbk/lightning/lnd/grpc/config/LndClientAutoConfigurationTest.java
@@ -5,20 +5,19 @@ import org.junit.jupiter.api.Test;
 import org.lightningj.lnd.wrapper.AsynchronousLndAPI;
 import org.springframework.beans.factory.NoSuchBeanDefinitionException;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
-import org.tbk.lightning.lnd.grpc.LndJsonRpcClientFactory;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class LndJsonRpcClientAutoConfigurationTest {
+public class LndClientAutoConfigurationTest {
 
     private final ApplicationContextRunner contextRunner = new ApplicationContextRunner();
 
     @Test
     public void beansAreCreated() {
-        this.contextRunner.withUserConfiguration(LndJsonRpcClientAutoConfiguration.class)
+        this.contextRunner.withUserConfiguration(LndClientAutoConfiguration.class)
                 .withPropertyValues(
                         "org.tbk.lightning.lnd.grpc.rpchost=localhost",
                         "org.tbk.lightning.lnd.grpc.rpcport=10001",
@@ -26,47 +25,26 @@ public class LndJsonRpcClientAutoConfigurationTest {
                         "org.tbk.lightning.lnd.grpc.certFilePath=src/test/resources/lnd/tls-test.cert"
                 )
                 .run(context -> {
-                    assertThat(context.containsBean("lndJsonRpcClientFactory"), is(true));
-                    assertThat(context.getBean(LndJsonRpcClientFactory.class), is(notNullValue()));
-
-                    assertThat(context.containsBean("lndJsonRpcClient"), is(true));
+                    assertThat(context.containsBean("lndClient"), is(true));
                     assertThat(context.getBean(AsynchronousLndAPI.class), is(notNullValue()));
                 });
     }
 
     @Test
     public void noBeansAreCreated() {
-        this.contextRunner.withUserConfiguration(LndJsonRpcClientAutoConfiguration.class)
+        this.contextRunner.withUserConfiguration(LndClientAutoConfiguration.class)
                 .withPropertyValues(
                         "org.tbk.lightning.lnd.grpc.enabled=false"
                 )
                 .run(context -> {
-                    assertThat(context.containsBean("lndJsonRpcClientFactory"), is(false));
-                    assertThrows(NoSuchBeanDefinitionException.class, () -> context.getBean(LndJsonRpcClientFactory.class));
-
-                    assertThat(context.containsBean("lndJsonRpcClient"), is(false));
-                    assertThrows(NoSuchBeanDefinitionException.class, () -> context.getBean(AsynchronousLndAPI.class));
-                });
-    }
-
-    @Test
-    public void onlyFactoryIsCreated() {
-        this.contextRunner.withUserConfiguration(LndJsonRpcClientAutoConfiguration.class)
-                .withPropertyValues(
-                        "org.tbk.lightning.lnd.grpc.enabled=true"
-                )
-                .run(context -> {
-                    assertThat(context.containsBean("lndJsonRpcClientFactory"), is(true));
-                    assertThat(context.getBean(LndJsonRpcClientFactory.class), is(notNullValue()));
-
-                    assertThat(context.containsBean("lndJsonRpcClient"), is(false));
+                    assertThat(context.containsBean("lndClient"), is(false));
                     assertThrows(NoSuchBeanDefinitionException.class, () -> context.getBean(AsynchronousLndAPI.class));
                 });
     }
 
     @Test
     public void errorIfCertFileIsMissing() {
-        this.contextRunner.withUserConfiguration(LndJsonRpcClientAutoConfiguration.class)
+        this.contextRunner.withUserConfiguration(LndClientAutoConfiguration.class)
                 .withPropertyValues(
                         "org.tbk.lightning.lnd.grpc.rpchost=localhost",
                         "org.tbk.lightning.lnd.grpc.rpcport=10001",