+++ changelog.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - spring-lnurl: Publish `LnurlAuthWalletActionEvent` on successfully authorized wallet requests
 
+### Removed
+- fix: 
+
 ### Changed
 - spring-lnurl: Improved customization support for `LnurlAuthConfigurer`
 - refactor: throw IllegalArgumentException instead of NPE on @NonNull violations
+- adapt class names in lnd client module (remove 'JsonRpc' string in names)
 - upgrade: update jmolecules from v2021.0.2 to v2021.1.0
 - upgrade: update consensusj from v0.5.8 to v0.5.9
 - upgrade: update xchange from v5.0.7 to v5.0.11